backend/requirements.txt
torch
torchvision
torchxrayvision
scikit-image
pandas
imageio
requests
fastapi
uvicorn[standard]
python-multipart
langchain
google-generativeai
python-dotenv
langchain-google-genai